Worked on the mockobjects
Made new class FlightExtend which extends the abstract Flight Class. Now
the flight group can make this FlightExtend class contain what ever the
like but it need to have at least the attributes of the Flight class. We
can then use the get methods from Flight to obtain the data. See the
main method in UserInterface for an example of usage.